Added some new packages for commands, added a reddit client framework stubbed out
