+++ pom.xml
@@ -98,6 +98,11 @@
             <artifactId>HikariCP</artifactId>
             <version>2.7.6</version>
         </dependency>
+        <dependency>
+            <groupId>net.dean.jraw</groupId>
+            <artifactId>JRAW</artifactId>
+            <version>1.0.0</version>
+        </dependency>
     </dependencies>
 
     <repositories>
