+++ src/main/java/com/bot/commands/voice/PauseCommand.java
@@ -1,28 +1,32 @@
-package com.bot.commands;
+package com.bot.commands.voice;
 
 import com.bot.voice.VoiceSendHandler;
 import com.jagrosh.jdautilities.commandclient.Command;
 import com.jagrosh.jdautilities.commandclient.CommandEvent;
 
-public class StopCommand extends Command {
+public class PauseCommand extends Command {
 
-	public StopCommand() {
-		this.name = "stop";
+	public PauseCommand() {
+		this.name = "pause";
 		this.arguments = "";
-		this.help = "Stops stream and clears the current playlist";
+		this.help = "Pauses or resumes the stream";
 	}
 
 	@Override
 	protected void execute(CommandEvent commandEvent) {
-		// TODO: Add more permissions checks
 		VoiceSendHandler handler = (VoiceSendHandler) commandEvent.getGuild().getAudioManager().getSendingHandler();
 		if (handler == null) {
 			commandEvent.reply(commandEvent.getClient().getWarning() + " I am not connected to a voice channel.");
 		}
 		else {
-			handler.stop();
-			commandEvent.reply(commandEvent.getClient().getSuccess() + " Stopped audio stream");
-			commandEvent.getGuild().getAudioManager().closeAudioConnection();
+			if (handler.getPlayer().isPaused()) {
+				handler.getPlayer().setPaused(false);
+				commandEvent.reply(commandEvent.getClient().getSuccess() + " Resumed stream.");
+			}
+			else {
+				handler.getPlayer().setPaused(true);
+				commandEvent.reply(commandEvent.getClient().getSuccess() + " Paused stream.");
+			}
 		}
 	}
 }
