+++ src/main/java/com/bot/commands/voice/PlayCommand.java
@@ -1,4 +1,4 @@
-package com.bot.commands;
+package com.bot.commands.voice;
 
 import com.bot.Bot;
 import com.bot.voice.QueuedAudioTrack;
