+++ src/main/java/com/bot/commands/voice/RemoveTrackCommand.java
@@ -1,16 +1,16 @@
-package com.bot.commands;
+package com.bot.commands.voice;
 
 import com.jagrosh.jdautilities.commandclient.Command;
 import com.jagrosh.jdautilities.commandclient.CommandEvent;
 
-public class SkipCommand extends Command {
+public class RemoveTrackCommand extends Command {
 
-	public SkipCommand() {
+	public RemoveTrackCommand() {
 		// TODO: init
 	}
 
 	@Override
 	protected void execute(CommandEvent commandEvent) {
-		// TODO: Find stream and start skip reaction to end it
+		// TODO: take track # and remove from current list
 	}
 }
