+++ src/main/java/com/bot/commands/voice/SavePlaylistCommand.java
@@ -1,16 +1,16 @@
-package com.bot.commands;
+package com.bot.commands.voice;
 
 import com.jagrosh.jdautilities.commandclient.Command;
 import com.jagrosh.jdautilities.commandclient.CommandEvent;
 
-public class RemovePlaylistCommand extends Command {
+public class SavePlaylistCommand extends Command {
 
-	public RemovePlaylistCommand() {
-		// TODO: Init
+	public SavePlaylistCommand() {
+		// TODO: init
 	}
 
 	@Override
 	protected void execute(CommandEvent commandEvent) {
-		// TODO: Remove from DB for user/guild?
+		// TODO: Take current playlist/range and save it to DB
 	}
 }
