+++ src/main/java/com/bot/commands/voice/StopCommand.java
@@ -1,32 +1,28 @@
-package com.bot.commands;
+package com.bot.commands.voice;
 
 import com.bot.voice.VoiceSendHandler;
 import com.jagrosh.jdautilities.commandclient.Command;
 import com.jagrosh.jdautilities.commandclient.CommandEvent;
 
-public class PauseCommand extends Command {
+public class StopCommand extends Command {
 
-	public PauseCommand() {
-		this.name = "pause";
+	public StopCommand() {
+		this.name = "stop";
 		this.arguments = "";
-		this.help = "Pauses or resumes the stream";
+		this.help = "Stops stream and clears the current playlist";
 	}
 
 	@Override
 	protected void execute(CommandEvent commandEvent) {
+		// TODO: Add more permissions checks
 		VoiceSendHandler handler = (VoiceSendHandler) commandEvent.getGuild().getAudioManager().getSendingHandler();
 		if (handler == null) {
 			commandEvent.reply(commandEvent.getClient().getWarning() + " I am not connected to a voice channel.");
 		}
 		else {
-			if (handler.getPlayer().isPaused()) {
-				handler.getPlayer().setPaused(false);
-				commandEvent.reply(commandEvent.getClient().getSuccess() + " Resumed stream.");
-			}
-			else {
-				handler.getPlayer().setPaused(true);
-				commandEvent.reply(commandEvent.getClient().getSuccess() + " Paused stream.");
-			}
+			handler.stop();
+			commandEvent.reply(commandEvent.getClient().getSuccess() + " Stopped audio stream");
+			commandEvent.getGuild().getAudioManager().closeAudioConnection();
 		}
 	}
 }
