+++ src/main/java/com/bot/commands/voice/VoiceStatsCommand.java
@@ -1,4 +1,4 @@
-package com.bot.commands;
+package com.bot.commands.voice;
 
 import com.bot.Bot;
 import com.jagrosh.jdautilities.commandclient.Command;
